initial refactor to ACTUALLY use the client.rs and atom.rs modules
